+++ src/entrypoint.sh
@@ -24,10 +24,16 @@ start_server() {
   exec gunicorn
 }
 
-start_celeryworker() {
-  log "Going to start celeryworker"
+start_celeryworker_jobs() {
+  log "Going to start celeryworker-jobs (queue: a-jobs)"
   mkdir -p /data/logs
-  exec celery --app=onprem worker --hostname=worker1@%n --logfile=/data/logs/%n-%i.log --statedb=/data/worker.state --loglevel="${SRE_CELERY_LOG_LEVEL:-INFO}" --without-heartbeat --without-gossip --without-mingle -Ofair --concurrency 1 --prefetch-multiplier 1
+  exec celery --app=onprem worker -Q a-jobs --hostname=worker-jobs@%n --logfile=/data/logs/%n-%i.log --statedb=/data/worker-jobs.state --loglevel="${SRE_CELERY_LOG_LEVEL:-INFO}" --without-heartbeat --without-gossip --without-mingle -Ofair --concurrency 1 --prefetch-multiplier 1
+}
+
+start_celeryworker_scheduled() {
+  log "Going to start celeryworker-scheduled (queue: b-scheduled)"
+  mkdir -p /data/logs
+  exec celery --app=onprem worker -Q b-scheduled --hostname=worker-scheduled@%n --logfile=/data/logs/%n-%i.log --statedb=/data/worker-scheduled.state --loglevel="${SRE_CELERY_LOG_LEVEL:-INFO}" --without-heartbeat --without-gossip --without-mingle -Ofair --concurrency 1 --prefetch-multiplier 1
 }
 
 start_celerybeat() {
@@ -40,8 +46,11 @@ case "$1" in
   celerybeat)
     start_celerybeat
     ;;
-  celeryworker)
-    start_celeryworker
+  celeryworker-jobs)
+    start_celeryworker_jobs
+    ;;
+  celeryworker-scheduled)
+    start_celeryworker_scheduled
     ;;
   *)
     start_server